+++ docs/news.md
@@ -13,9 +13,14 @@ These are new features and improvements of note in each release.
 
 ## mooda v0.1.0 (Work in progress)
 
-* In mooda/\_\_init\_\_.py: It does not load plotmap.py if you do not have installed the basemap library. We added a warning message.
+* In mooda/\_\_init\_\_.py: 
+  * It does not load plotmap.py if you do not have installed the basemap library. We added a warning message.
+  * \_\_version\_\_ added.
 * In mooda/waterframe.py:
   * barplot(): It creates the graph even without knowing the meaning of the parameters.
+  * \_\_repr\_\_: New method. Return a string containing a printable representation of an object.
+  * corr(): New function. Compute pairwise correlation of data columns, excluding NA/null values.
+  * max_diff(): New function. Calculation the maximum difference between values of two parameters.
 * In mooda/access/egim.py:
   * to_csv(): New function. It creates a CSV file following the OceanSites standard.